requirements.txt
mcp[cli]
azure-kusto-data
tabulate